+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Try wal truncate_frequency in alloy-metrics.
+
 ## [0.7.1] - 2024-10-10
 
 ### Fixed