+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add wal `truncate_frequency` configuration to alloy-metrics with a default set to 15m.
 - Add grafanaOrganization CRD in helm chart.
 
 ## [0.7.1] - 2024-10-10