+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add wal `truncate_frequency` configuration to alloy-metrics with a default set to 15m.
 - Add grafanaOrganization CRD in helm chart.
 
+### Changed
+
+- Change default default monitoring agent to Alloy
+
 ## [0.7.1] - 2024-10-10
 
 ### Fixed